+++ src/common/serialization/vst3.h
@@ -51,13 +51,19 @@
 
 /**
  * Marker struct to indicate the other side (the plugin) should send a copy of
- * the configuration.
+ * the configuration. During this process we will also transmit the version
+ * string from the host, so we can show a little warning when the user forgot to
+ * rerun `yabridgectl sync` (and the initialization was still successful).
  */
 struct WantsConfiguration {
     using Response = Configuration;
 
+    std::string host_version;
+
     template <typename S>
-    void serialize(S&) {}
+    void serialize(S& s) {
+        s.text1b(host_version, 128);
+    }
 };
 
 /**